Extract Topology
