+++ features/storage/src/main/kotlin/com/boswelja/truemanager/storage/overview/Topology.kt
@@ -1,0 +1,83 @@
+package com.boswelja.truemanager.storage.overview
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.ProvideTextStyle
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import com.boswelja.truemanager.core.api.v2.pool.Topology
+import com.boswelja.truemanager.core.api.v2.pool.VDev
+
+@Composable
+fun Topology(
+    topology: Topology,
+    modifier: Modifier = Modifier
+) {
+    Column(
+        modifier = modifier,
+        verticalArrangement = Arrangement.spacedBy(8.dp)
+    ) {
+        Text(
+            text = "Topology",
+            style = MaterialTheme.typography.titleMedium
+        )
+        TopologySection(
+            title = { Text("Data") },
+            vdevs = topology.data,
+            modifier = Modifier.fillMaxWidth()
+        )
+        TopologySection(
+            title = { Text("Cache") },
+            vdevs = topology.cache,
+            modifier = Modifier.fillMaxWidth()
+        )
+        TopologySection(
+            title = { Text("De-duplication") },
+            vdevs = topology.dedup,
+            modifier = Modifier.fillMaxWidth()
+        )
+        TopologySection(
+            title = { Text("Log") },
+            vdevs = topology.log,
+            modifier = Modifier.fillMaxWidth()
+        )
+        TopologySection(
+            title = { Text("Spare") },
+            vdevs = topology.spare,
+            modifier = Modifier.fillMaxWidth()
+        )
+        TopologySection(
+            title = { Text("Metadata") },
+            vdevs = topology.special,
+            modifier = Modifier.fillMaxWidth()
+        )
+    }
+}
+
+@Composable
+fun TopologySection(
+    title: @Composable () -> Unit,
+    vdevs: List<VDev>,
+    modifier: Modifier = Modifier
+) {
+    Column(modifier) {
+        ProvideTextStyle(MaterialTheme.typography.labelLarge) {
+            title()
+        }
+        if (vdevs.isNotEmpty()) {
+            vdevs.forEach { vdev ->
+                if (vdev.children.isNotEmpty()) {
+                    Text("• ${vdev.type} | ${vdev.children.count()} wide")
+                } else {
+                    Text("• ${vdev.type} | No redundancy")
+                }
+            }
+        } else {
+            Text("• VDEVs not assigned")
+        }
+    }
+}
